+++ pkg/middleware/middleware.go
@@ -27,6 +27,7 @@ type AxwaySubscription interface {
 	GetSubscriberEmailAddress() string
 	GetSubscriberUserName() string
 	GetRevisionName() string
+	GetRevisionId() string
 	IsEnvironmentDefined() bool
 	GetEnvironmentName() string
 	IsExternalAPIIDDefined() bool
@@ -167,6 +168,11 @@ func (c *SubscriptionContainer) GetSubscriberUserName() string {
 // GetRevisionName - Facade to retrieve RevisionName
 func (c *SubscriptionContainer) GetRevisionName() string {
 	return c.serviceRevision.GetName()
+}
+
+// GetRevisionId - Facade to retrieve RevisionId (metadata.id)
+func (c *SubscriptionContainer) GetRevisionId() string {
+	return c.serviceRevision.Metadata.ID
 }
 
 // IsEnvironmentDefined - Facade to check if environment is set in Service Instance
@@ -323,7 +329,7 @@ func (sm *SubscriptionMiddleware) ProcessUnsubscribeSubscription() error {
 
 // ProcessSubscription  - Orchestrates entire subscription steps
 func (sm *SubscriptionMiddleware) ProcessSubscription() error {
-	log.Infof("Provisioning Subscription triggered [Environment:%s] [Revision/API:%s][", sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName())
+	log.Infof("Provisioning Subscription triggered [Environment:%s] [Revision/API:%s] [RevisionId:%s]", sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName(), sm.AxSub.GetRevisionId())
 	provisionedAPI := false
 	provisionedAPIProduct := false
 	provisionedTeam := false
@@ -432,9 +438,9 @@ func (sm *SubscriptionMiddleware) ProcessSubscription() error {
 	}
 
 	dto := notification.SubscribeMetaDataDto{
-		API:             sm.AxSub.GetRevisionName(),
+		API:             sm.AxSub.GetRevisionName() + " (" + sm.AxSub.GetRevisionId() + ")",
 		Team:            sm.AxSub.GetSubscriptionOwningTeamID(),
-		Product:         sm.AxSub.GetRevisionName(),
+		Product:         sm.AxSub.GetRevisionName() + " (" + sm.AxSub.GetRevisionId() + ")",
 		Application:     sm.AxSub.GetSubscriptionID(),
 		Environment:     sm.AxSub.GetEnvironmentName(),
 		Subscription:    sm.AxSub.GetSubscriptionName(),
@@ -451,7 +457,7 @@ func (sm *SubscriptionMiddleware) ProcessSubscription() error {
 			log.Error("[ERROR] [MIDDLEWARE] [SUBSCRIBE] [Notification] [Notification was not sent successfully]")
 		}
 	}
-	log.Infof("Successfully provisioned Subscription [Environment:%s] [Revision/API:%s][", sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName())
+	log.Infof("Successfully provisioned Subscription [Environment:%s] [Revision/API:%s [RevisionId:%s][", sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName(), sm.AxSub.GetRevisionId())
 	return nil
 
 }
@@ -462,9 +468,9 @@ func (sm *SubscriptionMiddleware) NotifySuccess(trigger string, message string, 
 	username := sm.AxSub.GetSubscriberUserName()
 
 	dto := notification.MonitorDataDto{
-		API:             sm.AxSub.GetRevisionName(),
+		API:             sm.AxSub.GetRevisionName() + " (" + sm.AxSub.GetRevisionId() + ")",
 		Team:            sm.AxSub.GetSubscriptionOwningTeamID(),
-		Product:         sm.AxSub.GetRevisionName(),
+		Product:         sm.AxSub.GetRevisionName() + " (" + sm.AxSub.GetRevisionId() + ")",
 		Application:     sm.AxSub.GetSubscriptionID(),
 		Environment:     sm.AxSub.GetEnvironmentName(),
 		Subscription:    sm.AxSub.GetSubscriptionName(),
@@ -494,9 +500,9 @@ func (sm *SubscriptionMiddleware) NotifyFailure(trigger string, message string, 
 	username := sm.AxSub.GetSubscriberUserName()
 
 	dto := notification.MonitorDataDto{
-		API:             sm.AxSub.GetRevisionName(),
+		API:             sm.AxSub.GetRevisionName() + " (" + sm.AxSub.GetRevisionId() + ")",
 		Team:            sm.AxSub.GetSubscriptionOwningTeamID(),
-		Product:         sm.AxSub.GetRevisionName(),
+		Product:         sm.AxSub.GetRevisionName() + " (" + sm.AxSub.GetRevisionId() + ")",
 		Application:     sm.AxSub.GetSubscriptionID(),
 		Environment:     sm.AxSub.GetEnvironmentName(),
 		Subscription:    sm.AxSub.GetSubscriptionName(),
@@ -535,13 +541,13 @@ func (sm *SubscriptionMiddleware) IsEnvironmentAsOrgAvailable() (bool, error) {
 
 //IsConnectorAPIAvailable - Facade to check via Connector if API already exists
 func (sm *SubscriptionMiddleware) IsConnectorAPIAvailable() (bool, error) {
-	return connector.GetOrgConnector().IsAPIAvailable(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName())
+	return connector.GetOrgConnector().IsAPIAvailable(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionId())
 }
 
 //IsAPIProductAvailable - Facade to check via Connector if API-Product exists
 func (sm *SubscriptionMiddleware) IsAPIProductAvailable() (bool, error) {
 	//by convention api and product have same name: revisionName
-	return connector.GetOrgConnector().IsAPIProductAvailable(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName())
+	return connector.GetOrgConnector().IsAPIProductAvailable(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionId())
 }
 
 //IsConnectorTeamAvailable - Facade to check via Connector if Team exists
@@ -569,9 +575,25 @@ func (sm *SubscriptionMiddleware) PublishAPIProduct() error {
 	permissions := sm.AxSub.GetServiceAttributes()
 	for _, endpoint := range sm.AxSub.GetServiceInstanceSpecEndpoints() {
 
+		// get first connector environment with same host as axway environment host
 		idx := sort.Search(len(connectorEnvs), func(i int) bool {
 			return endpoint.Host == connectorEnvs[i].Host
 		})
+		if idx < len(connectorEnvs) && connectorEnvs[idx].Host == endpoint.Host {
+			envNames = append(envNames, connectorEnvs[idx].Name)
+
+			// add all protocols and their version to connector product - ignoring axway defined protocol
+			for protocol, version := range connectorEnvs[idx].ProtocolVersion {
+				ver := connector.CommonVersion(version)
+				protocols = append(protocols, connector.Protocol{
+					Name:    connector.ProtocolName(protocol),
+					Version: &ver})
+			}
+		} else {
+			return errors.New("Environment not found in Solace-Connector for Host:" + endpoint.Host)
+		}
+		// TODO introduce configuration option to enable strict environment protocol check
+		/**
 		if idx < len(connectorEnvs) && connectorEnvs[idx].Host == endpoint.Host {
 			envNames = append(envNames, connectorEnvs[idx].Name)
 			protocolVersion, found := connectorEnvs[idx].ProtocolVersion[endpoint.Protocol]
@@ -586,19 +608,20 @@ func (sm *SubscriptionMiddleware) PublishAPIProduct() error {
 		} else {
 			return errors.New("Environment not found")
 		}
+		*/
 	}
 
-	return connector.GetOrgConnector().PublishAPIProduct(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName(), []string{sm.AxSub.GetRevisionName()}, envNames, protocols, permissions)
+	return connector.GetOrgConnector().PublishAPIProduct(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionId(), []string{sm.AxSub.GetRevisionId()}, envNames, protocols, permissions)
 }
 
 //RemoveAPI - Facade to remove via connector an API
 func (sm *SubscriptionMiddleware) RemoveAPI(ignoreConflict bool) error {
-	return connector.GetOrgConnector().RemoveAPI(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName(), ignoreConflict)
+	return connector.GetOrgConnector().RemoveAPI(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionId(), ignoreConflict)
 }
 
 //RemoveAPIProduct - Facade to remove via Connector an API Product
 func (sm *SubscriptionMiddleware) RemoveAPIProduct(ignoreConflict bool) error {
-	return connector.GetOrgConnector().RemoveAPIProduct(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName(), ignoreConflict)
+	return connector.GetOrgConnector().RemoveAPIProduct(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionId(), ignoreConflict)
 }
 
 //GetTeamApp - Facade to retrieve App AxSub generic JSON
@@ -628,14 +651,14 @@ func (sm *SubscriptionMiddleware) GetAppApis() ([]*map[string]interface{}, error
 //RemoveTeamApp - Facade to remove via Connector a Team Application
 func (sm *SubscriptionMiddleware) RemoveTeamApp() error {
 	apiProducts := make([]string, 0)
-	apiProducts = append(apiProducts, sm.AxSub.GetRevisionName())
+	apiProducts = append(apiProducts, sm.AxSub.GetRevisionId())
 	return connector.GetOrgConnector().RemoveTeamApp(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetSubscriptionOwningTeamID(), sm.AxSub.GetSubscriptionID())
 }
 
 //PublishTeamApp - Facade to publish via Connector a Team Application
 func (sm *SubscriptionMiddleware) PublishTeamApp() (*connector.Credentials, error) {
 	apiProducts := make([]string, 0)
-	apiProducts = append(apiProducts, sm.AxSub.GetRevisionName())
+	apiProducts = append(apiProducts, sm.AxSub.GetRevisionId())
 	trustedCNSList := make([]string, 0)
 	var webHooks *connector.SolaceWebhook = nil
 	if len(sm.AxSub.GetSubscriptionPropertyValue(solace.SolaceHTTPMethod)) > 0 {
@@ -662,7 +685,7 @@ func (sm *SubscriptionMiddleware) PublishAPI() error {
 	if err != nil {
 		return err
 	}
-	return connector.GetOrgConnector().PublishAPI(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionName(), decodedAPISpec)
+	return connector.GetOrgConnector().PublishAPI(sm.AxSub.GetEnvironmentName(), sm.AxSub.GetRevisionId(), decodedAPISpec)
 }
 
 //PublishTeam - Facade to publish via Connector a Team